changes last week

+++ app/main/requests.py
@@ -7,33 +7,36 @@ from .. import db
 from ..models import Request
 
 
-@main.route('/api/v1/requests', methods=['POST'])
+@main.route('/api/v1/requests/', methods=['POST'])
 @auth.login_required
 def create_request():
-    """ create request
+    """ Makes a new meetup request
+    Only logged in users can make meetup requests. The id of the maker of the
+    request is stored in each request object.
     :return: return JSON of new request
     """
     content = request.get_json(force=True)
     location_string = content.get('location_string')
     if not location_string:
-        return 'Location required'
+        return jsonify(message='Location required')
     latitude, longitude = get_geocode_location(location_string)
     meal_request = Request(
+        user_id=g.user.id,
         meal_type=content.get('meal_type'),
         location_string=location_string,
-        latitude=latitude, longitude=longitude,
+        latitude=latitude,
+        longitude=longitude,
         meal_time=content.get('meal_time'),
-        user_id=g.user.id,
-        filled=content.get('filled'), )
+        filled=content.get('filled'))
     db.session.add(meal_request)
     db.session.commit()
     return jsonify(meal_request.serialize)
 
 
-@main.route('/api/v1/requests')
+@main.route('/api/v1/requests/')
 @auth.login_required
 def get_all_requests():
-    """ show all open requests
+    """ Shows all open meetup requests.
     :return: return JSON of all requests
     """
     user = g.user
@@ -48,7 +51,7 @@ def get_all_requests():
 @main.route('/api/v1/requests/<int:id>')
 @auth.login_required
 def get_request(id):
-    """ show request with given id
+    """ show information for specific meetup request
     :param id: request id (int)
     :return: return JSON of request
     """
@@ -61,14 +64,19 @@ def get_request(id):
 @main.route('/api/v1/requests/<int:id>', methods=['PUT'])
 @auth.login_required
 def update_request(id):
-    """ change request with given id
+    """ Updates information about a meetup request.
+    Only the original maker of the request should be able to edit it.
     :param id: request id (int)
     :return: return JSON of changed request
     """
+    user = g.user
     content = request.get_json(force=True)
     meal_request = Request.get_record_by_id(id)
     if not meal_request:
         return jsonify(message='Meal request not found')
+    if user.id != meal_request.user_id:
+        return jsonify(message='Only the original maker of the request can '
+                               'edit it.')
     updated_meal_request = meal_request.update(content)
     if not updated_meal_request:
         return jsonify(message='Nothing to update')
@@ -79,11 +87,16 @@ def update_request(id):
 @auth.login_required
 def delete_request(id):
     """ delete request with given id
+    Only the original maker of the request should be able to delete it.
     :param id: request id (int)
     :return: return JSON with message that request has been deleted
     """
+    user = g.user
     meal_request = Request.get_record_by_id(id)
     if not meal_request:
         return jsonify(message='Meal request not found')
+    if user.id != meal_request.user_id:
+        return jsonify(message='Only the original maker of the request can '
+                               'delete it.')
     meal_request.delete()
     return jsonify(message='Meal request has been deleted')
